app/requirements.txt
flask
requests
python-dotenv
os